Add files via upload
DS3002 Final Project.dbc contains the notebook which has the code showing the whole process. I also included the excel spreadsheet with my PivotChart and a screenshot of it as well. Query1-4,1-5,1-6 JSON files are the ones I used for the streaming process which filtered some of the Sales Order Details table by 3 batches of Product ID. The Kaggle-2 JSON file contains my key so that the API can be used. Query 1 is the JSON file I used to populate my MongoDB collection.